+++ Sources/CMapboxCoreNavigation/CMapboxCoreNavigation.h
@@ -6,4 +6,4 @@ FOUNDATION_EXPORT double CMapboxCoreNavigationVersionNumber;
 //! Project version string for CMapboxCoreNavigation.
 FOUNDATION_EXPORT const unsigned char CMapboxCoreNavigationVersionString[];
 
-#import "MBXAccounts+CoreNavigationAdditions.h"
+#import "MBXAccounts.h"