+++ Sources/CMapboxCoreNavigation/include/MBXAccounts+CoreNavigationAdditions.h
@@ -1,9 +1,0 @@
-#import <MapboxAccounts/MapboxAccounts.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface MBXAccounts (CoreNavigationAdditions)
-
-@end
-
-NS_ASSUME_NONNULL_END