+++ Sources/CMapboxCoreNavigation/include/MBXAccounts.h
@@ -1,0 +1,11 @@
+#import <MapboxCommon/MapboxCommon.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MBXAccounts : NSObject
+
+@property (class, nonatomic, readonly, nullable) NSString *serviceSkuToken;
+
+@end
+
+NS_ASSUME_NONNULL_END